Separated some cogs (Nerds and Queries)
