+++ cogs/__init__.py
@@ -1,11 +1,10 @@
 import discord
 from dotenv import load_dotenv
 from json import dump, load
-import math
 import os
 from random import randint
 from rdoclient import RandomOrgClient
-from typing import Callable, Dict, Union
+from typing import Dict, Union
 
 # Discourse
 def load_cfg(name: str, default_dict: Dict={}) -> Dict:
@@ -21,29 +20,9 @@ def load_cfg(name: str, default_dict: Dict={}) -> Dict:
 
 _cmd=load_cfg('commands')
 _resp=load_cfg('responses')
-_wiki=load_cfg('wiki')
 _global=load_cfg('global')
 _guild: Dict[int, Dict[str, int]]={}
 
-_math_func_dict: Dict[str, Callable]={
-    'ceil': math.ceil, 'comb': math.comb, 'copysign': math.copysign,
-    'abs': math.fabs, 'factorial': math.factorial, 'floor': math.floor,
-    'fmod': math.fmod, 'frexp': math.frexp, 'fsum': math.fsum, 'gcd': math.gcd,
-    'isclose': math.isclose, 'isfinite': math.isfinite, 'isinf': math.isinf,
-    'isnan': math.isnan, 'isqrt': math.isqrt, 'ldexp': math.ldexp,
-    'modf': math.modf, 'perm': math.perm, 'prod': math.prod,
-    'remainder': math.remainder, 'trunc': math.trunc, 'exp': math.exp,
-    'expm1': math.expm1, 'log': math.log, 'log1p': math.log1p,
-    'log2': math.log2, 'log10': math.log10, 'pow': math.pow, 'sqrt': math.sqrt,
-    'acos': math.acos, 'asin': math.asin, 'atan': math.atan,
-    'atan2': math.atan2, 'cos': math.cos, 'dist': math.dist,
-    'hypot': math.hypot, 'sin': math.sin, 'tan': math.tan,
-    'degrees': math.degrees, 'radians': math.radians, 'acosh': math.acosh,
-    'asinh': math.asinh, 'atanh': math.atanh, 'cosh': math.cosh,
-    'sinh': math.sinh, 'tanh': math.tanh, 'erf': math.erf, 'erfc': math.erfc,
-    'gamma': math.gamma, 'lgamma': math.lgamma, 'pi': math.pi, 'e': math.e,
-    'tau': math.tau, 'inf': math.inf, 'nan': math.nan
-}
 # Randorg
 load_dotenv()
 _rdo: RandomOrgClient=RandomOrgClient(os.getenv('RANDORG_API_KEY'))
@@ -58,7 +37,7 @@ def get_emoji(guild: discord.Guild, name: str, fallback_id: int=None) -> Union[
         return f'<:{name}:{fallback_id}>'
     return emoji
 
-def get_guild_cfg(guild: discord.Guild) -> None:
+def load_guild_cfg(guild: discord.Guild) -> None:
     global _guild
 
     if not os.path.exists('data/guilds/'):
@@ -71,3 +50,9 @@ def get_guild_cfg(guild: discord.Guild) -> None:
         'lorekeep': 0,
         'mute': 0
     })
+
+def get_possessive(self, noun: str) -> str:
+    if noun[-1]=='s':
+        return noun+'\''
+    else:
+        return noun+'\'s'
